+++ demisto_sdk/commands/common/tests/tools_test.py
@@ -22,7 +22,10 @@ from demisto_sdk.commands.common.tools import (LOG_COLORS,
                                                get_latest_release_notes_text,
                                                get_matching_regex,
                                                get_release_notes_file_path,
-                                               get_ryaml, retrieve_file_ending,
+                                               get_ryaml,
+                                               has_remote_configured,
+                                               is_origin_content_repo,
+                                               retrieve_file_ending,
                                                run_command_os,
                                                server_version_compare)
 from demisto_sdk.tests.constants_test import (INDICATORFIELD_EXTRA_FIELDS,
@@ -344,3 +347,49 @@ def test_get_release_notes_file_path_invalid():
     """
     filepath = '/SomePack/1_1_1.json'
     assert get_release_notes_file_path(filepath) is None
+
+
+remote_testbank = [
+    ('origin  https://github.com/turbodog/content.git', False),
+    ('upstream  https://github.com/demisto/content.git', True)
+]
+
+
+@pytest.mark.parametrize('git_value, response', remote_testbank)
+def test_has_remote(mocker, git_value, response):
+    """
+    While: Testing if the remote upstream contains demisto/content
+    Given:
+      1. Origin string not containing demisto/content
+      2. Upstream string containing demisto/content
+    Expects:
+      1. Test condition fails
+      2. Test condition passes
+    :param git_value: Git string from `git remotes -v`
+    """
+    mocker.patch('demisto_sdk.commands.common.tools.run_command', return_value=git_value)
+    test_remote = has_remote_configured()
+    assert response == test_remote
+
+
+origin_testbank = [
+    ('origin  https://github.com/turbodog/content.git', False),
+    ('origin  https://github.com/demisto/content.git', True)
+]
+
+
+@pytest.mark.parametrize('git_value, response', origin_testbank)
+def test_origin_content(mocker, git_value, response):
+    """
+    While: Testing if the remote origin contains demisto/content
+    Given:
+      1. Origin string not containing demisto/content
+      2. Origin string containing demisto/content
+    Expects:
+      1. Test condition fails
+      2. Test condition passes
+    :param git_value: Git string from `git remotes -v`
+    """
+    mocker.patch('demisto_sdk.commands.common.tools.run_command', return_value=git_value)
+    test_remote = is_origin_content_repo()
+    assert response == test_remote