+++ demisto_sdk/commands/common/update_id_set.py
@@ -47,6 +47,42 @@ CHECKED_TYPES_REGEXES = (
     PACKS_MAPPER_JSON_REGEX
 )
 
+BUILT_IN_FIELDS = [
+    "name",
+    "details",
+    "severity",
+    "owner",
+    "dbotCreatedBy",
+    "type",
+    "dbotSource",
+    "category",
+    "dbotStatus",
+    "playbookId",
+    "dbotCreated",
+    "dbotClosed",
+    "occurred",
+    "dbotDueDate",
+    "dbotModified",
+    "dbotTotalTime",
+    "reason",
+    "closeReason",
+    "closeNotes",
+    "closingUserId",
+    "reminder",
+    "phase",
+    "roles",
+    "labels",
+    "attachment",
+    "runStatus",
+    "sourceBrand",
+    "sourceInstance",
+    "CustomFields",
+    "droppedCount",
+    "linkedCount",
+    "feedBased",
+    "id"
+]
+
 
 def checked_type(file_path, regex_list=CHECKED_TYPES_REGEXES):
     for regex in regex_list:
@@ -168,11 +204,19 @@ def get_integration_data(file_path):
     cmd_list = [command.get('name') for command in commands]
     pack = get_pack_name(file_path)
     integration_api_modules = get_integration_api_modules(file_path, data_dictionary, is_unified_integration)
+    default_classifier = data_dictionary.get('defaultClassifier')
+    default_incident_type = data_dictionary.get('defaultIncidentType')
+    is_feed = data_dictionary.get('feed')
+    mappers = set()
 
     deprecated_commands = []
     for command in commands:
         if command.get('deprecated', False):
             deprecated_commands.append(command.get('name'))
+
+    for mapper in ['defaultMapperIn', 'defaultMapperOut']:
+        if data_dictionary.get(mapper):
+            mappers.add(data_dictionary.get(mapper))
 
     integration_data['name'] = name
     integration_data['file_path'] = file_path
@@ -192,7 +236,131 @@ def get_integration_data(file_path):
         integration_data['pack'] = pack
     if integration_api_modules:
         integration_data['api_modules'] = integration_api_modules
+    if default_classifier and default_classifier != '':
+        integration_data['classifiers'] = default_classifier
+    if mappers:
+        integration_data['mappers'] = list(mappers)
+    if default_incident_type and default_incident_type != '':
+        integration_data['incident_types'] = default_incident_type
+    if is_feed:
+        integration_data['indicator_fields'] = "CommonTypes"
+        integration_data['indicator_types'] = "CommonTypes"
+
     return {id_: integration_data}
+
+
+def get_fields_by_script_argument(task):
+    """Iterates over the task script arguments and search for non empty fields
+
+    Args:
+        task (dict): A task of the playbook with `script: Builtin|||setIncident`
+
+    Returns:
+        set. set of incident fields related to this task
+
+    Example:
+        for the task:
+            {
+                task:
+                  id: e80e3f5a-a74f-44a8-8e83-8eee96def1d0
+                  name: Save authenticity check result to incident field
+                  script: Builtin|||setIncident
+
+                scriptarguments:
+                  emailaddress:
+                    complex:
+                      root: ActiveDirectory
+                      accessor: Users.mail
+                  duration: {}
+            }
+
+        we will return the 'emailaddress` incident field
+    """
+
+    dependent_incident_fields = set()
+    for field_name, field_value in task.get('scriptarguments', {}).items():
+        if field_value and field_name not in BUILT_IN_FIELDS:
+            if field_name != "customFields":
+                dependent_incident_fields.add(field_name)
+            else:
+                # the value should be a list of dicts in str format
+                custom_field_value = list(field_value.values())[0]
+                if isinstance(custom_field_value, str):
+                    custom_fields_list = json.loads(custom_field_value)
+                    for custom_field in custom_fields_list:
+                        field_name = list(custom_field.keys())[0]
+                        if field_name not in BUILT_IN_FIELDS:
+                            dependent_incident_fields.add(field_name)
+    return dependent_incident_fields
+
+
+def get_incident_fields_by_playbook_input(input):
+    """Searches for incident fields in a playbook input.
+
+    Args:
+        input (dict): An input of the playbook
+
+    Returns:
+        set. set of incident fields related to this task
+    """
+    dependent_incident_fields = set()
+
+    input_type = list(input.keys())[0]  # type can be `simple` or `complex`
+    input_value = list(input.values())[0]
+
+    # check if it is in the form 'simple: ${incident.field_name}'
+    if input_type == 'simple' and str(input_value).startswith('${incident.'):
+        field_name = input_value.split('.')[1][:-1]
+        if field_name not in BUILT_IN_FIELDS:
+            dependent_incident_fields.add(field_name)
+
+    elif input_type == 'complex':
+        root_value = str(input_value.get('root', ''))
+        accessor_value = str(input_value.get('accessor'))
+        combined_value = root_value + '.' + accessor_value  # concatenate the strings
+
+        field_name = re.match(r'incident\.([^\.]+)', combined_value)
+        if field_name:
+            field_name = field_name.groups()[0]
+            if field_name not in BUILT_IN_FIELDS:
+                dependent_incident_fields.add(field_name)
+
+    return dependent_incident_fields
+
+
+def get_dependent_incident_and_indicator_fields(data_dictionary):
+    """Finds the incident fields and indicator fields dependent on this playbook
+
+    Args:
+        data_dictionary (dict): The playbook data dict
+
+    Returns:
+        set. set of incident fields related to this playbook
+    """
+    dependent_incident_fields = set()
+    dependent_indicator_fields = set()
+    for task in data_dictionary.get('tasks', {}).values():
+        # incident fields dependent by field mapping
+        related_incident_fields = task.get('fieldMapping')
+        if related_incident_fields:
+            for incident_field in related_incident_fields:
+                if incident_field not in BUILT_IN_FIELDS:
+                    dependent_incident_fields.add(incident_field.get('incidentfield'))
+
+        # incident fields dependent by scripts arguments
+        if 'setIncident' in task.get('task', {}).get('script', ''):
+            dependent_incident_fields.update(get_fields_by_script_argument(task))
+            # incident fields dependent by scripts arguments
+        if 'setIndicator' in task.get('task', {}).get('script', ''):
+            dependent_indicator_fields.update(get_fields_by_script_argument(task))
+
+    # incident fields by playbook inputs
+    for input in data_dictionary.get('inputs', []):
+        input_value_dict = input.get('value', {})
+        if input_value_dict and isinstance(input_value_dict, dict):  # deprecated playbooks bug
+            dependent_incident_fields.update(get_incident_fields_by_playbook_input(input_value_dict))
+
+    return dependent_incident_fields, dependent_indicator_fields
 
 
 def get_playbook_data(file_path: str) -> dict:
@@ -212,6 +380,7 @@ def get_playbook_data(file_path: str) -> dict:
     skippable_tasks = (implementing_scripts_skippable + implementing_playbooks_skippable +
                        command_to_integration_skippable)
     pack = get_pack_name(file_path)
+    dependent_incident_fields, dependent_indicator_fields = get_dependent_incident_and_indicator_fields(data_dictionary)
 
     playbook_data['name'] = name
     playbook_data['file_path'] = file_path
@@ -234,7 +403,10 @@ def get_playbook_data(file_path: str) -> dict:
         playbook_data['pack'] = pack
     if skippable_tasks:
         playbook_data['skippable_tasks'] = skippable_tasks
-
+    if dependent_incident_fields:
+        playbook_data['incident_fields'] = list(dependent_incident_fields)
+    if dependent_indicator_fields:
+        playbook_data['indicator_fields'] = list(dependent_indicator_fields)
     return {id_: playbook_data}
 
 
@@ -416,6 +588,71 @@ def get_incident_field_data(path, incidents_types_list):
         data['incident_types'] = list(all_associated_types)
     if all_scripts:
         data['scripts'] = list(all_scripts)
+
+    return {id_: data}
+
+
+def get_indicator_type_data(path):
+    data = OrderedDict()
+    json_data = get_json(path)
+
+    id_ = json_data.get('id')
+    name = json_data.get('details', '')
+    fromversion = json_data.get('fromVersion')
+    toversion = json_data.get('toVersion')
+    associated_integration = json_data.get('reputationCommand')
+    pack = get_pack_name(path)
+    all_scripts = set()
+
+    for field in ['reputationScriptName', 'enhancementScriptNames']:
+        associated_scripts = json_data.get(field)
+        if associated_scripts == '':
+            continue
+
+        associated_scripts = [associated_scripts] if not isinstance(associated_scripts, list) else associated_scripts
+        if associated_scripts:
+            all_scripts = all_scripts.union(set(associated_scripts))
+
+    if name:
+        data['name'] = name
+    if toversion:
+        data['toversion'] = toversion
+    if fromversion:
+        data['fromversion'] = fromversion
+    if pack:
+        data['pack'] = pack
+    if associated_integration:
+        data['integrations'] = associated_integration
+    if all_scripts:
+        data['scripts'] = list(all_scripts)
+
+    return {id_: data}
+
+
+def get_incident_type_data(path):
+    data = OrderedDict()
+    json_data = get_json(path)
+
+    id_ = json_data.get('id')
+    name = json_data.get('name', '')
+    fromversion = json_data.get('fromVersion')
+    toversion = json_data.get('toVersion')
+    playbook_id = json_data.get('playbookId')
+    pre_processing_script = json_data.get('preProcessingScript')
+    pack = get_pack_name(path)
+
+    if name:
+        data['name'] = name
+    if toversion:
+        data['toversion'] = toversion
+    if fromversion:
+        data['fromversion'] = fromversion
+    if pack:
+        data['pack'] = pack
+    if playbook_id and playbook_id != '':
+        data['playbooks'] = playbook_id
+    if pre_processing_script and pre_processing_script != '':
+        data['scripts'] = pre_processing_script
 
     return {id_: data}
 
@@ -695,7 +932,7 @@ def process_incident_types(file_path: str, print_logs: bool) -> list:
     if checked_type(file_path, [PACKS_INCIDENT_TYPE_JSON_REGEX]):
         if print_logs:
             print("adding {} to id_set".format(file_path))
-        res.append(get_general_data(file_path))
+        res.append(get_incident_type_data(file_path))
     return res
 
 
@@ -733,7 +970,7 @@ def process_indicator_types(file_path: str, print_logs: bool) -> list:
             not checked_type(file_path, [INDICATOR_TYPES_REPUTATIONS_REGEX, PACKS_INDICATOR_TYPES_REPUTATIONS_REGEX])):
         if print_logs:
             print("adding {} to id_set".format(file_path))
-        res.append(get_general_data(file_path))
+        res.append(get_indicator_type_data(file_path))
     return res