+++ demisto_sdk/commands/create_id_set/tests/id_set_creator_test.py
@@ -5,8 +5,11 @@ from tempfile import mkdtemp
 import pytest
 from demisto_sdk.commands.common.git_tools import git_path
 from demisto_sdk.commands.common.update_id_set import (
-    get_classifier_data, get_incident_field_data, get_layout_data,
-    get_mapper_data, get_values_for_keys_recursively, has_duplicate)
+    get_classifier_data, get_fields_by_script_argument,
+    get_incident_field_data, get_incident_fields_by_playbook_input,
+    get_incident_type_data, get_indicator_type_data, get_layout_data,
+    get_mapper_data, get_playbook_data, get_values_for_keys_recursively,
+    has_duplicate)
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 
 
@@ -165,14 +168,102 @@ def test_get_incident_fields_data_no_types_scripts():
     Then
         - parsing all the data from file successfully
     """
-    test_dir = \
-        f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidentfield-to-test-no-types_scripts.json'
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidentfield-to-test-no-types_scripts.json'
     result = get_incident_field_data(test_dir, [])
     result = result.get('incidentfield-test')
     assert 'name' in result.keys()
     assert 'fromversion' in result.keys()
     assert 'toversion' in result.keys()
     assert 'incident_types' not in result.keys()
+    assert 'scripts' not in result.keys()
+
+
+def test_get_indicator_type_data():
+    """
+    Given
+        - An indicator type file called reputation-indicatortype.json.
+
+    When
+        - parsing indicator type files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/reputation-indicatortype.json'
+
+    result = get_indicator_type_data(test_dir)
+    result = result.get('indicator-type-dummy')
+    assert 'name' in result.keys()
+    assert 'fromversion' in result.keys()
+    assert 'integrations' in result.keys()
+    assert 'scripts' in result.keys()
+    assert "dummy-script" in result.get('scripts')
+    assert "dummy-script-2" in result.get('scripts')
+    assert "dummy-script-3" in result.get('scripts')
+
+
+def test_get_indicator_type_data_no_integration_no_scripts():
+    """
+    Given
+        - An indicator type file called reputation-indicatortype_no_script_no_integration.json without any
+            integrations or scripts that it depends on.
+
+    When
+        - parsing indicator type files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/'
+    f'reputation-indicatortype_no_script_no_integration.json'
+
+    result = get_indicator_type_data(test_dir)
+    result = result.get('indicator-type-dummy')
+    assert 'name' in result.keys()
+    assert 'fromversion' in result.keys()
+    assert 'integrations' not in result.keys()
+    assert 'scripts' not in result.keys()
+
+
+def test_get_incident_types_data():
+    """
+    Given
+        - An incident type file called incidenttype-to-test.json
+
+    When
+        - parsing incident type files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidenttype-to-test.json'
+    result = get_incident_type_data(test_dir)
+    result = result.get('dummy incident type')
+    assert 'name' in result.keys()
+    assert 'fromversion' in result.keys()
+    assert 'playbooks' in result.keys()
+    assert 'scripts' in result.keys()
+
+
+def test_get_incident_types_data_no_playbooks_scripts():
+    """
+    Given
+        - An incident type file called incidenttype-to-test-no-playbook-script.json with no script or playbook
+        related to it
+
+    When
+        - parsing incident type files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/incidenttype-to-test-no-playbook-script.json'
+
+    result = get_incident_type_data(test_dir)
+    result = result.get('dummy incident type')
+    assert 'name' in result.keys()
+    assert 'fromversion' in result.keys()
+    assert 'playbooks' not in result.keys()
     assert 'scripts' not in result.keys()
 
 
@@ -210,8 +301,8 @@ def test_get_classifiers_data_no_types_scripts():
     Then
         - parsing all the data from file successfully
     """
-    test_dir = \
-        f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-to-test-no-incidenttypes.json'
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-to-test-no-incidenttypes.json'
+
     result = get_classifier_data(test_dir)
     result = result.get('dummy classifier')
     assert 'name' in result.keys()
@@ -259,8 +350,8 @@ def test_get_mappers_data_no_types_fields():
     Then
         - parsing all the data from file successfully
     """
-    test_dir = \
-        f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-mapper-to-test-no-types-fields.json'
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/classifier-mapper-to-test-no-types-fields.json'
+
     result = get_mapper_data(test_dir)
     result = result.get('dummy mapper')
     assert 'name' in result.keys()
@@ -310,3 +401,171 @@ def test_get_values_for_keys_recursively():
     }
 
     assert expected == get_values_for_keys_recursively(test_dict, test_keys)
+
+
+def test_get_playbook_data():
+    """
+    Given
+        - A playbook file called playbook-with-incident-fields.yml
+
+    When
+        - parsing playbook files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-with-incident-fields.yml'
+    result = get_playbook_data(test_dir)
+    result = result.get('Arcsight - Get events related to the Case')
+    assert 'name' in result.keys()
+    assert 'file_path' in result.keys()
+    assert 'implementing_scripts' in result.keys()
+    assert 'implementing_scripts' in result.keys()
+    assert 'command_to_integration' in result.keys()
+    assert 'tests' in result.keys()
+    assert 'incident_fields' in result.keys()
+    assert 'indicator_fields' in result.keys()
+
+
+def test_get_playbook_data_no_fields():
+    """
+    Given
+        - A playbook file called playbook-no-incident-fields.yml without any
+            incident or indicator fields that it depends on.
+
+    When
+        - parsing playbook files
+
+    Then
+        - parsing all the data from file successfully
+    """
+    test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-no-incident-fields.yml'
+    result = get_playbook_data(test_dir)
+    result = result.get('Arcsight - Get events related to the Case')
+    assert 'name' in result.keys()
+    assert 'file_path' in result.keys()
+    assert 'implementing_scripts' in result.keys()
+    assert 'implementing_scripts' in result.keys()
+    assert 'command_to_integration' in result.keys()
+    assert 'tests' in result.keys()
+    assert 'incident_fields' not in result.keys()
+    assert 'indicator_fields' not in result.keys()
+
+
+INPUT_WITH_INCIDENT_FIELD_SIMPLE = {
+    "key": "AlertID",
+    "value": {
+        "simple": "${incident.field_name}"
+    },
+    "required": False
+}
+
+INPUT_WITH_INCIDENT_FIELD_COMPLEX1 = {
+    "key": "AlertID",
+    "value": {
+        "complex": {
+            "root": "incident",
+            "accessor": "field_name"
+        }
+    },
+    "required": False
+}
+
+INPUT_WITH_INCIDENT_FIELD_COMPLEX2 = {
+    "key": "AlertID",
+    "value": {
+        "complex": {
+            "root": "incident.field_name",
+            "accessor": "username"
+        }
+    },
+    "required": False
+}
+
+INPUT_SIMPLE_WITHOUT_INCIDENT_FIELD = {
+    "key": "AlertID",
+    "value": {
+        "simple": "${not_incident.field_name}"
+    },
+    "required": False
+}
+
+INPUT_COMPLEX_WITHOUT_INCIDENT_FIELD = {
+    "key": "AlertID",
+    "value": {
+        "complex": {
+            "root": "something",
+            "accessor": "username"
+        }
+    },
+    "required": False
+}
+
+INPUTS = [
+    (INPUT_WITH_INCIDENT_FIELD_SIMPLE, True),
+    (INPUT_WITH_INCIDENT_FIELD_COMPLEX1, True),
+    (INPUT_WITH_INCIDENT_FIELD_COMPLEX2, True),
+    (INPUT_SIMPLE_WITHOUT_INCIDENT_FIELD, False),
+    (INPUT_COMPLEX_WITHOUT_INCIDENT_FIELD, False)
+]
+
+
+@pytest.mark.parametrize('playbook_input, are_there_incident_fields', INPUTS)
+def test_get_incident_fields_by_playbook_input(playbook_input, are_there_incident_fields):
+    """
+    Given
+        - A list of playbook inputs
+
+    When
+        - Searching for dependent incident fields
+
+    Then
+        -  Finding all dependent incident fields in the input
+    """
+
+    result = get_incident_fields_by_playbook_input(input=playbook_input.get('value'))
+    if are_there_incident_fields:
+        assert "field_name" in result
+    else:
+        assert result == set()
+
+
+EXAMPLE_TASK_WITH_SIMPLE_SCRIPT_ARGUMENTS = {
+    "id": "ID",
+    "scriptarguments": {
+        "field_name": {
+            "simple": "${inputs.IndicatorTagName}"
+        }
+    }
+}
+
+EXAMPLE_TASK_WITH_CUSTOM_FIELDS_SCRIPT_ARGUMENTS = {
+    "id": "ID",
+    "scriptarguments": {
+        "customFields": {
+            "simple": '[{"field_name":"${inputs.IndicatorTagName}"}]'
+        }
+    }
+}
+
+TASK_INPUTS = [
+    # EXAMPLE_TASK_WITH_SIMPLE_SCRIPT_ARGUMENTS,
+    EXAMPLE_TASK_WITH_CUSTOM_FIELDS_SCRIPT_ARGUMENTS
+]
+
+
+@pytest.mark.parametrize('task', TASK_INPUTS)
+def test_get_fields_by_script_argument(task):
+    """
+    Given
+        - A list of playbook tasks
+
+    When
+        - Searching for dependent incident fields in the task script arguments
+
+    Then
+        - Finding all dependent incident fields in the task
+    """
+
+    result = get_fields_by_script_argument(task)
+    assert "field_name" in result