+++ demisto_sdk/commands/find_dependencies/tests/find_dependencies_test.py
@@ -379,6 +379,37 @@ class TestDependsOnPlaybook:
         for found_result in found_result_set:
             assert not found_result[1]  # validate that mandatory is set to False
 
+    def test_collect_playbooks_dependencies_on_incident_fields(self, id_set):
+        expected_result = {("DigitalGuardian", True), ("EmployeeOffboarding", True)}
+        test_input = [
+            {
+                "Dummy Playbook": {
+                    "name": "Dummy Playbook",
+                    "file_path": "dummy_path",
+                    "fromversion": "dummy_version",
+                    "implementing_scripts": [
+                    ],
+                    "implementing_playbooks": [
+                    ],
+                    "command_to_integration": {
+                    },
+                    "tests": [
+                        "dummy_playbook"
+                    ],
+                    "pack": "dummy_pack",
+                    "incident_fields": [
+                        "digitalguardianusername",
+                        "Google Display Name"
+                    ]
+                }
+            }
+        ]
+
+        found_result = PackDependencies._collect_playbooks_dependencies(pack_playbooks=test_input,
+                                                                        id_set=id_set)
+
+        assert found_result == expected_result
+
 
 class TestDependsOnLayout:
     @pytest.mark.parametrize("dependency_types, dependency_fields ,expected_result",
@@ -387,6 +418,16 @@ class TestDependsOnLayout:
                               ("Fake", "Fake", set())
                               ])
     def test_collect_layouts_dependencies(self, dependency_types, dependency_fields, expected_result, id_set):
+        """
+        Given
+            - A layout entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the layout depends on.
+        """
         test_input = [
             {
                 "Dummy Layout": {
@@ -416,7 +457,17 @@ class TestDependsOnIncidentField:
                               ("Fake", "Fake", set()),
                               ("Fake", "Fake", set())
                               ])
-    def test_collect_layouts_dependencies(self, dependency_types, dependency_scripts, expected_result, id_set):
+    def test_collect_incident_field_dependencies(self, dependency_types, dependency_scripts, expected_result, id_set):
+        """
+        Given
+            - An incident field entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the incident field depends on.
+        """
         test_input = [
             {
                 "Dummy Incident Field": {
@@ -434,6 +485,118 @@ class TestDependsOnIncidentField:
         ]
         found_result = PackDependencies._collect_incidents_fields_dependencies(
             pack_incidents_fields=test_input, id_set=id_set)
+
+        # TODO: update the test once the implementation of all dependencies is working
+        assert found_result == expected_result
+
+
+class TestDependsOnIndicatorType:
+    @pytest.mark.parametrize("dependency_integrations, dependency_scripts ,expected_result",
+                             [("Fake", "Fake", set()),
+                              ("Fake", "Fake", set()),
+                              ("Fake", "Fake", set())
+                              ])
+    def test_collect_layouts_dependencies(self, dependency_integrations, dependency_scripts, expected_result, id_set):
+        """
+        Given
+            - An indicator type entry in the id_set.
+
+        When
+            - Building dependency graph for pack.
+
+        Then
+            - Extracting the packs that the indicator type depends on.
+        """
+        test_input = [
+            {
+                "Dummy Indicator Type": {
+                    "name": "Dummy Indicator Type",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "integrations": [
+                        dependency_integrations
+                    ],
+                    "scripts": [
+                        dependency_scripts
+                    ]
+                }
+            }
+        ]
+        found_result = PackDependencies._collect_indicators_types_dependencies(
+            pack_indicators_types=test_input, id_set=id_set)
+
+        # TODO: update the test once the implementation of all dependencies is working
+        assert found_result == expected_result
+
+
+class TestDependsOnIntegrations:
+    @pytest.mark.parametrize(
+        "dependency_classifiers, dependency_mappers, dependency_incident_types, dependency_indicator_fields, "
+        "dependency_indicator_types, expected_result",
+        [("Fake", [], "Fake", "Fake", "Fake", set()),
+         ("Fake", [], "Fake", "Fake", "Fake", set()),
+         ("Fake", [], "Fake", "Fake", "Fake", set())
+         ])
+    def test_collect_layouts_dependencies(self, dependency_classifiers, dependency_mappers, dependency_incident_types,
+                                          dependency_indicator_fields, dependency_indicator_types, expected_result,
+                                          id_set):
+        """
+        Given
+            - An integration entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the integration depends on.
+        """
+        test_input = [
+            {
+                "Dummy Integration": {
+                    "name": "Dummy Incident Field",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "classifiers": dependency_classifiers,
+                    "mappers": dependency_mappers,
+                    "incident_types": dependency_incident_types,
+                    "indicator_fields": dependency_indicator_fields,
+                    "indicator_types": dependency_indicator_types
+                }
+            }
+        ]
+        found_result = PackDependencies._collect_incidents_fields_dependencies(
+            pack_incidents_fields=test_input, id_set=id_set)
+
+        # TODO: update the test once the implementation of all dependencies is working
+        assert found_result == expected_result
+
+
+class TestDependsOnIncidentType:
+    @pytest.mark.parametrize("dependency_playbook, dependency_scripts ,expected_result",
+                             [("Fake", "Fake", set()),
+                              ("Fake", "Fake", set()),
+                              ("Fake", "Fake", set())
+                              ])
+    def test_collect_incident_type_dependencies(self, dependency_playbook, dependency_scripts, expected_result, id_set):
+        """
+        Given
+            - An incident type entry in the id_set.
+        When
+            - Building dependency graph for pack.
+        Then
+            - Extracting the packs that the incident type depends on.
+        """
+        test_input = [
+            {
+                "Dummy Incident Type": {
+                    "name": "Dummy Incident Type",
+                    "fromversion": "5.0.0",
+                    "pack": "dummy_pack",
+                    "playbooks": dependency_playbook,
+                    "scripts": dependency_scripts
+                }
+            }
+        ]
+        found_result = PackDependencies._collect_incidents_types_dependencies(
+            pack_incidents_types=test_input, id_set=id_set)
 
         # TODO: update the test once the implementation of all dependencies is working
         assert found_result == expected_result